Add images of primary sources to some tweets
Using the random module, the bot will now periodically crop an image of
a nineteenth-century primary source (e.g., advertisement, bill of sale)
related to slave sales and include it in the tweet, along with permalink
to the source online.

The current images are listed in a CSV spreadsheet that also includes
named persons, whose names are also periodically included in these image
tweets.

Thanks to @samplereality, @wynkenhimself, @CharleneCac and others for
suggestions about adding images.

https://twitter.com/wcaleb/status/539532062345015296